# static/swagger.yml
openapi: 3.0.0
info:
  title: Data Forest API
  description: Data Forest API
  version: 0.0.1

servers:
  - url: http://localhost:5000/
    description: Local server

paths:
  /users:
    get:
      summary: List users
      description: Get a paginated list of users.
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      summary: Create a user
      description: Creates a new user with the given details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input data
        "409":
          description: Email is already in use

  /users/{user_id}:
    get:
      summary: Get a user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found

    put:
      summary: Update a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input data
        "404":
          description: User not found
        "409":
          description: Email is already in use

    delete:
      summary: Delete a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User deleted successfully
        "404":
          description: User not found
  /reforested_areas:
    get:
      summary: List reforested areas
      description: Get a paginated list of reforested areas.
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        "200":
          description: A list of reforested areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReforestedArea"
    post:
      summary: Create a reforested area
      description: Creates a new reforested area with the given details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReforestedAreaInput"
      responses:
        "201":
          description: Reforested area created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReforestedArea"
        "400":
          description: Invalid input data
  /reforested_areas/{area_id}:
    get:
      summary: Get a reforested area by ID
      parameters:
        - name: area_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reforested area retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReforestedArea"
        "404":
          description: Reforested area not found

    put:
      summary: Update a reforested area
      parameters:
        - name: area_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReforestedAreaInput"
      responses:
        "200":
          description: Reforested area updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReforestedArea"
        "400":
          description: Invalid input data
        "404":
          description: Reforested area not found

    delete:
      summary: Delete a reforested area
      parameters:
        - name: area_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reforested area deleted successfully
        "404":
          description: Reforested area not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - full_name
        - email

    UserInput:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - full_name
        - email
        - password
    ReforestedArea:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        area_in_m2:
          type: number
          format: float
        geom:
          type: object
          properties:
            type:
              type: string
            coordinates:
              type: array
              items:
                type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - name
        - description
        - area_in_m2
        - geom
        - created_at
        - updated_at
    ReforestedAreaInput:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        area_in_m2:
          type: number
          format: float
        geom:
          type: object
          properties:
            type:
              type: string
            coordinates:
              type: array
              items:
                type: number
      required:
        - user_id
        - name
        - description
        - area_in_m2
        - geom